Extract default eChart options
